bug fixed

+++ models_doc.py
@@ -203,7 +203,7 @@ def calc_coeff(iter_num, high=1.0, low=0.0, alpha=10.0, max_iter=10000.0):
     Returns:
         float: A calculated coefficient in the range [low, high].
     """
-    return np.float(2.0 * (high - low) / (1.0 + np.exp(-alpha * iter_num / max_iter)) - (high - low) + low)
+    return np.float64(2.0 * (high - low) / (1.0 + np.exp(-alpha * iter_num / max_iter)) - (high - low) + low)
 
 def grl_hook(coeff):
     """
